requirements.txt
gps_time
requests
python-dotenv